requirements.txt
numpy>=1.10.4
pyyaml
tornado
pyzmq
pyjwt
